# src/main/resources/openapi/api-docs.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0.0.1"
  },
  "paths": {
    "/runSync": {
      "post": {
        "tags": [
          "track-time-controller"
        ],
        "summary": "Вызвать синхронный метод.",
        "operationId": "runSync",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/runSync/{countOfCalls}": {
      "post": {
        "tags": [
          "track-time-controller"
        ],
        "summary": "Вызвать синхронный метод N раз.",
        "operationId": "runSyncNTimes",
        "parameters": [
          {
            "name": "countOfCalls",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/runAsync": {
      "post": {
        "tags": [
          "track-time-controller"
        ],
        "summary": "Вызвать асинхронный метод.",
        "operationId": "runAsync",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/runAsync/{countOfCalls}": {
      "post": {
        "tags": [
          "track-time-controller"
        ],
        "summary": "Вызвать асинхронный метод N раз.",
        "operationId": "runAsyncNTimes",
        "parameters": [
          {
            "name": "countOfCalls",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/findAll": {
      "get": {
        "tags": [
          "track-time-controller"
        ],
        "summary": "Получить статистику по всем ранее запущенным методам.",
        "operationId": "findByAll",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MethodExecutionStatisticsDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/find/{methodName}": {
      "get": {
        "tags": [
          "track-time-controller"
        ],
        "summary": "Получить статистику метода по имени.",
        "operationId": "findByName",
        "parameters": [
          {
            "name": "methodName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MethodExecutionStatisticsDto"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "MethodExecutionStatisticsDto": {
        "type": "object",
        "properties": {
          "methodName": {
            "type": "string"
          },
          "countOfCalls": {
            "type": "integer",
            "format": "int32"
          },
          "lastCallDateTime": {
            "type": "string",
            "format": "date-time"
          },
          "lastExecutionTime": {
            "type": "integer",
            "format": "int64"
          },
          "avgExecutionTime": {
            "type": "number",
            "format": "double"
          }
        }
      }
    }
  }
}